fix: `coc-java` doesn't work
* The error message TypeError: Cannot read properties of null (reading '0')

Adding the version number of lombok to the actual lombok jar filename. For
example changing "lombok.jar" to "lombok-1.18.30.jar". Make sure to follow
to format precisely.

* The running environment doesn't satisfy accordingly

The tooling JDK will be used to launch the Language Server for Java. And by
default, will also be used to compile your projects. Java 17 is the minimum
required version.