fix: cargo compilation errors
In `cargo install <package> --locked` means using the version required in
Cargo.lock, which is the version that was used when the lock file was last
updated. This is useful to ensure that the same version is used in different
environments, and to avoid the need to update the version manually.

Refs: https://stackoverflow.com/questions/76989793/why-cargo-install-needs-locked-to-work